--- src/library/DIPS.Mobile.UI/Effects/ContextMenuEffect/ContextMenuPlatformEffect.cs
@@ -1,10 +1,0 @@
-using Microsoft.Maui.Controls.Platform;
-
-namespace DIPS.Mobile.UI.Effects.ContextMenuEffect;
-
-public partial class ContextMenuPlatformEffect : PlatformEffect
-{
-    protected override partial void OnAttached();
-    protected override partial void OnDetached();
-
-}
